feat: Add asynchronous stack list loader for improved page load performance
